📈 perf: domain 별로 testContainer 를 하나씩만 가지도록 수정
