+++ src/test/kotlin/pmeet/pmeetserver/project/ProjectCommentIntegrationTest.kt
@@ -157,19 +157,4 @@ internal class ProjectCommentIntegrationTest : DescribeSpec() {
       }
     }
   }
-
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
 }
