+++ src/test/kotlin/pmeet/pmeetserver/project/ProjectIntegrationTest.kt
@@ -985,19 +985,4 @@ internal class ProjectIntegrationTest : DescribeSpec() {
       }
     }
   }
-
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
 }
