+++ src/test/kotlin/pmeet/pmeetserver/project/ProjectTryoutIntegrationTest.kt
@@ -110,19 +110,4 @@ internal class ProjectTryoutIntegrationTest : DescribeSpec() {
       }
     }
   }
-
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
 }
