+++ src/test/kotlin/pmeet/pmeetserver/project/repository/ProjectRepositoryUnitTest.kt
@@ -408,19 +408,4 @@ internal class ProjectRepositoryUnitTest(
     }
   }
 
-}) {
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
-}
+})
