+++ src/test/kotlin/pmeet/pmeetserver/project/repository/ProjectTryoutRepositoryUnitTest.kt
@@ -12,8 +12,6 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest
 import org.springframework.data.mongodb.core.ReactiveMongoTemplate
 import org.springframework.data.mongodb.repository.support.ReactiveMongoRepositoryFactory
-import org.testcontainers.containers.MongoDBContainer
-import org.testcontainers.junit.jupiter.Container
 import pmeet.pmeetserver.project.domain.ProjectTryout
 import pmeet.pmeetserver.project.domain.enum.ProjectTryoutStatus
 import java.time.LocalDateTime
@@ -21,7 +19,7 @@ import java.time.LocalDateTime
 @ExperimentalCoroutinesApi
 @DataMongoTest
 internal class ProjectTryoutRepositoryUnitTest(
-  @Autowired private val template: ReactiveMongoTemplate
+  @Autowired private val template: ReactiveMongoTemplate,
 ) : DescribeSpec({
 
   isolationMode = IsolationMode.InstancePerLeaf
@@ -35,6 +33,7 @@ internal class ProjectTryoutRepositoryUnitTest(
   lateinit var projectTryout: ProjectTryout
 
   beforeSpec {
+
     Dispatchers.setMain(testDispatcher)
 
     projectTryout = ProjectTryout(
@@ -65,19 +64,4 @@ internal class ProjectTryoutRepositoryUnitTest(
       }
     }
   }
-}) {
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
-}
+})
