+++ src/test/kotlin/pmeet/pmeetserver/user/UserIntegrationTest.kt
@@ -35,20 +35,6 @@ import org.springframework.test.context.ActiveProfiles
 @ExperimentalCoroutinesApi
 @ActiveProfiles("test")
 internal class UserIntegrationTest : DescribeSpec() {
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
 
   @Autowired
   lateinit var webTestClient: WebTestClient
