+++ src/test/kotlin/pmeet/pmeetserver/user/job/JobIntegrationTest.kt
@@ -133,20 +133,4 @@ internal class JobIntegrationTest : DescribeSpec() {
       }
     }
   }
-
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
-
 }
