+++ src/test/kotlin/pmeet/pmeetserver/user/job/repository/JobRepositoryUnitTest.kt
@@ -100,20 +100,5 @@ internal class JobRepositoryUnitTest(
     }
   }
 
-}) {
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
-}
+})
