+++ src/test/kotlin/pmeet/pmeetserver/user/repository/UserRepositoryUnitTest.kt
@@ -154,19 +154,4 @@ internal class UserRepositoryUnitTest(
     }
   }
 
-}) {
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
-}
+})
