+++ src/test/kotlin/pmeet/pmeetserver/user/techStack/TechStackIntegrationTest.kt
@@ -134,20 +134,4 @@ class TechStackIntegrationTest : DescribeSpec() {
       }
     }
   }
-
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
-  
 }
