+++ src/test/kotlin/pmeet/pmeetserver/user/techStack/repository/TechStackRepositoryUnitTest.kt
@@ -100,20 +100,5 @@ internal class TechStackRepositoryUnitTest(
     }
   }
 
-}) {
-  companion object {
-    @Container
-    val mongoDBContainer = MongoDBContainer("mongo:latest").apply {
-      withExposedPorts(27017)
-      start()
-    }
-
-    init {
-      System.setProperty(
-        "spring.data.mongodb.uri",
-        "mongodb://localhost:${mongoDBContainer.getMappedPort(27017)}/test"
-      )
-    }
-  }
-}
+})
